% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/greenclust.R
\name{greenclust}
\alias{greenclust}
\title{Row Clustering Using Greenacre's Method}
\usage{
greenclust(x, correct = FALSE, verbose = FALSE)
}
\arguments{
\item{x}{a numeric matrix or data frame}

\item{correct}{a logical indicating whether to apply a continuity
correction if and when the clustered table reaches a 2x2 dimension.}

\item{verbose}{if TRUE, prints the clustered table along with r-squared and
p-value at each step}
}
\value{
An object of class \code{greenclust} which is compatible with most
  most \code{\link{hclust}} object functions, such as \code{\link{plot}()} and
  \code{\link{rect.hclust}()}
}
\description{
Iteratively collapses the rows of a table (typically a contingency table)
by selecting the pair of rows each time whose combination results in a
new table with the smallest loss of chi-squared.
}
\examples{
# Combine Titanic passenger attributes into a single category
tab <- t(as.data.frame(apply(Titanic, 4:1, FUN=sum)))
# Remove rows with all zeros
tab <- tab[apply(tab, 1, sum) > 0, ]

# Perform clustering and plot
gc <- greenclust(tab)
plot(gc)

# Get suggested clusters and add to plot
clusters <- greencut(gc)
rect.hclust(gc, max(clusters))

# Plot r-squared and p-values for each cut point
greenplot(gc)
}
