% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assigncluster.R
\name{assign.cluster}
\alias{assign.cluster}
\title{Assign clusters to a new vector of categories}
\usage{
assign.cluster(x, clusters)
}
\arguments{
\item{x}{a factor or character vector representing a categorical variable}

\item{clusters}{a numeric vector representing cluster numbers, such as an
object returned by \code{\link{greenclust}} or \code{\link{hclust}}}

\item{src}{optional factor or character vector of categories to use for
the cluster numbers in \code{cluster} if they do not already exist in
the  \code{names} attribute of  \code{cluster}}
}
\description{
Maps a vector of cluster numbers to another categorical vector,
yielding a new vector of matching cluster numbers.
}
\details{
Useful for cases where clustering was performed on a table of unique levels
(as with \code{\link{greenclust}}), and the clusters need to be distributed
back out to the original observations. In this case, the \code{cluster}
vector will typically have a \code{names} attribute of corresponding
categories, and \code{src} will not be used.

Also handy for assigning clusters to a test data set when the original
clustering was done on a training set. Here, the names of \code{cluster}
will represent original observations rather than category names, so the
\code{src} vector will be needed.

Any categories in \code{x} that do not exist as either names in
\code{cluster} or elements in \code{src} are assigned the most frequent
cluster number (with ties going to the lowest value).
}
\examples{
# Example goes here

}
